% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{create_importance_matrix}
\alias{create_importance_matrix}
\title{Create Importance Matrix for Single Response Models}
\usage{
create_importance_matrix(imp)
}
\arguments{
\item{imp}{A list containing the importance data generated by \code{importance.preGroup}.}
}
\value{
A matrix with variable names as columns and types of importance as rows (linear, prognostic, prescriptive).
}
\description{
Generates a matrix of importances for each variable based on the importances calculated in \code{importance.preGroup}
for models with a single response variable.
}
\details{
This function organizes importances into a matrix format for easier visualization and analysis,
distinguishing between different types of importances.
}
\examples{
# Assuming `imp` is the list returned from `importance.preGroup`
importance_matrix <- create_importance_matrix(imp)
}
