% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{get_new_X}
\alias{get_new_X}
\title{Generate Model Matrix from New Data for preGroup Model}
\usage{
get_new_X(preGroup_model, new_data)
}
\arguments{
\item{preGroup_model}{A \code{preGroup} model object, specifically its \code{pre_fit} component.}

\item{new_data}{New data to be transformed into model matrix format.}
}
\value{
A list containing the transformed new data as a model matrix.
}
\description{
This function processes new data to fit the structure of a \code{preGroup} model,
facilitating the prediction process.
}
\details{
The function retrieves settings and parameters from the \code{pre_fit} component of
the \code{preGroup} model and applies these to the new data to generate a suitable model matrix.
This involves applying transformation rules and handling different types of predictors as
specified in the model.
}
\examples{
# Assuming `model` is a preGroup model and `new_data` is ready
mod_matrix <- get_new_X(model, new_data)
}
