% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{importance.preGroup}
\alias{importance.preGroup}
\alias{importance}
\title{Calculate Variable Importances for a preGroup Model}
\usage{
\method{importance}{preGroup}(
  preGroup_model,
  standardize = FALSE,
  global = TRUE,
  penalty.par.val = "lambda.1se",
  gamma = NULL,
  quantprobs = c(0.75, 1),
  round = NA,
  plot = TRUE,
  ylab = "Importance",
  main = "Variable importances",
  abbreviate = 10L,
  diag.xlab = TRUE,
  diag.xlab.hor = 0,
  diag.xlab.vert = 2,
  cex.axis = 1,
  legend = "topright",
  ...
)
}
\arguments{
\item{preGroup_model}{An object of class 'preGroup'.}

\item{standardize}{Logical; if TRUE, importance scores are standardized.}

\item{global}{Logical; if TRUE, global importances are calculated, otherwise local importances are provided.}

\item{penalty.par.val}{A character string specifying the penalty parameter value to be used in calculation.}

\item{gamma}{The gamma parameter for regularization paths in models; if not NULL, it modifies the penalties.}

\item{quantprobs}{A numeric vector specifying the quantiles to be used in local importance calculation.}

\item{round}{Numeric; the number of decimal places to round the importances to (if NA, no rounding is performed).}

\item{plot}{Logical; if TRUE, a plot of the importances is generated.}

\item{ylab}{The y-axis label in the plot.}

\item{main}{The main title for the plot.}

\item{abbreviate}{An integer specifying how much to abbreviate variable names in the plot.}

\item{diag.xlab}{Logical; if TRUE, labels are diagonal on the x-axis of the plot.}

\item{diag.xlab.hor}{Horizontal adjustment for diagonal labels.}

\item{diag.xlab.vert}{Vertical adjustment for diagonal labels.}

\item{cex.axis}{The character expansion size for axis names.}

\item{legend}{A character string or logical; position of the legend in the plot or FALSE to exclude it.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list containing the variable importances as data frames.
}
\description{
Computes variable importances based on a preGroup model, offering options for both global and local importance, 
standardization of importances, and various additional parameters.
}
\examples{
result <- importance(preGroup_model)
}
