% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{specify_imps}
\alias{specify_imps}
\title{Specify linear, prognostic, and prescriptive importances}
\usage{
specify_imps(imp_in)
}
\arguments{
\item{imp}{A list containing the importance data generated by `importance.preGroup`.}
}
\value{
A list: Element varimps is a matrix with importance type (linear, prognostic, prescriptive) as columns and variable names as rows. Element baseimps is a data frame containing rule importances and variable importances in each rule.
}
\description{
Generates a matrix of importances for each variable based on the importances calculated in `importance.preGroup` 
for models with a single response variable.
}
\details{
This function
}
