% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r, R/preTreatment.r
\name{summary.preTreatment}
\alias{summary.preTreatment}
\title{Summarize a preTreatment Model}
\usage{
\method{summary}{preTreatment}(object)

\method{summary}{preTreatment}(object)
}
\arguments{
\item{object}{An object of class 'preTreatment'.}
}
\value{
A summary list of the preTreatment model. Can extract the the tree generating algorithm, variable selection algorithm. and number of nonzero coefficients

A summary list of the preTreatment model. Can extract the the tree generating algorithm, variable selection algorithm. and number of nonzero coefficients
}
\description{
Provides a summary of the preTreatment model, including the tree generating algorithm, variable selection algorithm, and the number of nonzero coefficients for linear, prognostic, and prescriptive predictors.

Provides a summary of the preTreatment model, including the tree generating algorithm, variable selection algorithm, and the number of nonzero coefficients for linear, prognostic, and prescriptive predictors.
}
