% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{create_importance_matrix_multivariate}
\alias{create_importance_matrix_multivariate}
\title{Create Importance Matrices for Multivariate Response Models}
\usage{
create_importance_matrix_multivariate(imp)
}
\arguments{
\item{imp}{A list containing the importance data generated by \code{importance.preGroup}.}
}
\value{
A list of matrices, each corresponding to a different response variable in the model.
}
\description{
Generates a list of matrices of importances for each variable and each response in a multivariate model
based on the importances calculated in \code{importance.preGroup}.
}
\details{
Each matrix in the list organizes importances for a specific response, formatted similarly to \code{create_importance_matrix}.
}
\examples{
# Assuming `imp` is the list returned from `importance.preGroup`
importance_matrices <- create_importance_matrix_multivariate(imp)
}
