% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r, R/preTreatment.r
\name{get_new_X}
\alias{get_new_X}
\title{Generate Model Matrix from New Data for preTreatment Model}
\usage{
get_new_X(preTreatment_model, new_data)

get_new_X(preTreatment_model, new_data)
}
\arguments{
\item{preTreatment_model}{A \code{preTreatment} model object, specifically its \code{pre_fit} component.}

\item{new_data}{New data to be transformed into model matrix format.}
}
\value{
A list containing the transformed new data as a model matrix.

A list containing the transformed new data as a model matrix.
}
\description{
This function processes new data to fit the structure of a \code{preTreatment} model,
facilitating the prediction process.

This function processes new data to fit the structure of a \code{preTreatment} model,
facilitating the prediction process.
}
\details{
The function retrieves settings and parameters from the \code{pre_fit} component of
the \code{preTreatment} model and applies these to the new data to generate a suitable model matrix.
This involves applying transformation rules and handling different types of predictors as
specified in the model.

The function retrieves settings and parameters from the \code{pre_fit} component of
the \code{preTreatment} model and applies these to the new data to generate a suitable model matrix.
This involves applying transformation rules and handling different types of predictors as
specified in the model.
}
\examples{
# Assuming `model` is a preTreatment model and `new_data` is ready
\dontrun{
mod_matrix <- get_new_X(model, new_data)
}
# Assuming `model` is a preTreatment model and `new_data` is ready
\dontrun{
mod_matrix <- get_new_X(model, new_data)
}
}
