% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{predict.preGroup}
\alias{predict.preGroup}
\title{Predict Responses Using preGroup Model}
\usage{
\method{predict}{preGroup}(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{A `preGroup` model object from which predictions will be made.}

\item{newdata}{New data on which predictions are to be made.}

\item{type}{The type of prediction to be made: 'response' or 'HTE'. Type 'response' returns the predicted 
response values, while 'HTE' returns the heterogeneous treatment effects (for causal modelling).}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A vector of predicted responses.
}
\description{
Predicts responses based on a `preGroup` model and new data.
}
\details{
The function checks if the model object is a `preGroup` and then uses the
`get_new_X` function to transform new data into the model matrix format. Predictions
are then made using the multivariate structure of the `preGroup` model.
}
\examples{
# Assuming `model` is a preGroup model and `new_data` is available
\dontrun{
mod <- preGroup(Species ~ ., data = iris, treatment_indicator = "Petal.Width")
new_data <- iris[1:5, ]
predictions <- predict(mod, new_data)
}
}
