% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r, R/preTreatment.r
\name{specify_imps}
\alias{specify_imps}
\title{Specify linear, prognostic, and prescriptive importances}
\usage{
specify_imps(imp_in)

specify_imps(imp_in)
}
\arguments{
\item{imp_in}{A list containing the importance data generated by \code{importance.preTreatment}.}
}
\value{
A list: Element varimps is a matrix with importance type (linear, prognostic, prescriptive) as columns and variable names as rows. Element baseimps is a data frame containing rule importances and variable importances in each rule.

A list: Element varimps is a matrix with importance type (linear, prognostic, prescriptive) as columns and variable names as rows. Element baseimps is a data frame containing rule importances and variable importances in each rule.
}
\description{
Generates a matrix of importances for each variable based on the importances calculated in \code{importance.preTreatment}
for models with a single response variable.

Generates a matrix of importances for each variable based on the importances calculated in \code{importance.preTreatment}
for models with a single response variable.
}
\details{
This function only supports univariate outcome. Multivariate version needs to be further implemented.

This function only supports univariate outcome. Multivariate version needs to be further implemented.
}
