% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGroup.r
\name{summary.preGroup}
\alias{summary.preGroup}
\title{Summarize a preGroup Model}
\usage{
\method{summary}{preGroup}(object)
}
\arguments{
\item{object}{An object of class 'preGroup'.}
}
\value{
A summary list of the preGroup model. Can extract the the tree generating algorithm, variable selection algorithm. and number of nonzero coefficients
}
\description{
Provides a summary of the preGroup model, including the tree generating algorithm, variable selection algorithm, and the number of nonzero coefficients for linear, prognostic, and prescriptive predictors.
}
